Split Post helpers into extensions
